KrzYmod: TRUNK
A lot of changes, that's why for now marked as trunk until enough tests
are done.
- Master Block size increased to 100MB,
- read maximum of Master Block size data to memory then save that Master
Block compressed data to reduce memory usage on very big files while
keeping last output buffer char for bit updating,
- Make Zopfli Cache more intelligent by mallocing less memory on big
blocks (there is a control value of 512MB that checks predicted mallocs
for cache, however, cache can't be lower than 3*blocksize so this
control value may be exceeded at times),
- If out of memory occurs during cache malloc try to malloc lowest
possible value for char buffer, if it fails exit zopfli,
- Read files >2GB up to 4GB for 32 bit versions and bigger than that for
64 bit versions (may be of use for ZLIB or RAW Deflate, however, GZIP
doesn't specifically disallow input files >4GB - filesize in archive
will be reported incorrectly but it should still be decompressable,
tests are required),
- Make simple prediction that input file could exceed 4GB ZIP limit:
error in single file compress mode OR skip file in multi file compress
mode,
- Clean cbs/cbd switches data silently on files >100MB as they will not
work properly then,
- Use temp file output_filename.zopfli and ask for confirmation to
overwrite if output_filename exists (done after compression finishes).

Due to these changes (notably using less memory for input/output
operations) some radical changes were required like:
- moved ZIP/GZ time support to bin passable through 4 byte buffer
pointer to lib while, in case of ZIP output, that buffer gets updated to
8 bytes of crc and compressed size to be properly altered at FilePK
offset by bin - if NULL is passed as a pointer then write zero bytes to
gzip and oldest possible date to ZIP (however, in case of ZIP NULL may
cause lib to not return crc and compressed size to be updated in FilePK
header, rendering compressed file unreadable),
- some other small changes of less importance.

As far as ZopfliDeflate changes require 2 additional parameters since
v14, 0 and NULL respectively can be passed instead of full data size and
data position. Also passing >ZOPFLI_MASTER_BLOCK_SIZE data to
ZopfliCompress should make Zopfli work as usual.
This was also made to keep zopflipng compatibility without much changes.

As mentioned, TRUNK changes are not thoroughly tested so crashes or
undefined behaviors may occur. So far it was tested a bit on windows OS,
I'm not sure if code portability is kept the same as before and if it
works on other OSes. If I notice something, I will most likely fix it.

+++ src/zopfli/lz77.c
@@ -95,7 +95,7 @@ void ZopfliVerifyLenDist(const unsigned char* data, size_t datasize, size_t pos,
                          unsigned short dist, unsigned short length) {
 
   /* TODO(lode): make this only run in a debug compile, it's for assert only. */
-  size_t i;
+  unsigned short i;
 
   assert(pos + length <= datasize);
   for (i = 0; i < length; i++) {
